# perf(html/parser): atom + perf