--- a/crates/swc_html_ast/src/base.rs
+++ b/crates/swc_html_ast/src/base.rs
@@ -1,6 +1,6 @@
 use is_macro::Is;
 use string_enum::StringEnum;
-use swc_atoms::JsWord;
+use swc_atoms::{Atom, JsWord};
 use swc_common::{ast_node, EqIgnoreSpan, Span};
 
 #[ast_node("Document")]
@@ -64,7 +64,7 @@ pub struct DocumentType {
     #[cfg_attr(feature = "rkyv", with(swc_atoms::EncodeJsWord))]
     pub system_id: Option<JsWord>,
     #[cfg_attr(feature = "rkyv", with(swc_atoms::EncodeJsWord))]
-    pub raw: Option<JsWord>,
+    pub raw: Option<Atom>,
 }
 
 impl EqIgnoreSpan for DocumentType {
@@ -127,11 +127,11 @@ pub struct Attribute {
     #[cfg_attr(feature = "rkyv", with(swc_atoms::EncodeJsWord))]
     pub name: JsWord,
     #[cfg_attr(feature = "rkyv", with(swc_atoms::EncodeJsWord))]
-    pub raw_name: Option<JsWord>,
+    pub raw_name: Option<Atom>,
     #[cfg_attr(feature = "rkyv", with(swc_atoms::EncodeJsWord))]
     pub value: Option<JsWord>,
     #[cfg_attr(feature = "rkyv", with(swc_atoms::EncodeJsWord))]
-    pub raw_value: Option<JsWord>,
+    pub raw_value: Option<Atom>,
 }
 
 impl EqIgnoreSpan for Attribute {
@@ -150,7 +150,7 @@ pub struct Text {
     #[cfg_attr(feature = "rkyv", with(swc_atoms::EncodeJsWord))]
     pub data: JsWord,
     #[cfg_attr(feature = "rkyv", with(swc_atoms::EncodeJsWord))]
-    pub raw: Option<JsWord>,
+    pub raw: Option<Atom>,
 }
 
 impl EqIgnoreSpan for Text {
@@ -166,7 +166,7 @@ pub struct Comment {
     #[cfg_attr(feature = "rkyv", with(swc_atoms::EncodeJsWord))]
     pub data: JsWord,
     #[cfg_attr(feature = "rkyv", with(swc_atoms::EncodeJsWord))]
-    pub raw: Option<JsWord>,
+    pub raw: Option<Atom>,
 }
 
 impl EqIgnoreSpan for Comment {