<!-- Note: CI script will automatically rebase your PR so please do not rebase unless required -->

<!-- Thanks for submitting a pull request! Please provide enough information so that others can review your pull request. -->

**Description:**

Before using `swc_html_minify` to bench:
```
html/minify/document/css_spec                                                                           
                        time:   [48.143 ms 48.257 ms 48.390 ms]
                        change: [-3.1896% -0.9565% +0.6793%] (p = 0.46 > 0.05)
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high severe
html/minify/document/github                                                                           
                        time:   [19.532 ms 19.569 ms 19.618 ms]
                        change: [-3.1896% -0.9565% +0.6793%] (p = 0.46 > 0.05)
                        No change in performance detected.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high severe
html/minify/document/stackoverflow                                                                           
                        time:   [16.958 ms 17.008 ms 17.081 ms]
                        change: [-1.1692% -0.3571% +0.3620%] (p = 0.43 > 0.05)
                        No change in performance detected.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild

html/minify/document_fragment/css_spec                                                                           
                        time:   [46.740 ms 46.826 ms 47.056 ms]
                        change: [-0.7255% +0.1777% +1.0953%] (p = 0.73 > 0.05)
                        No change in performance detected.
Found 2 outliers among 10 measurements (20.00%)
  2 (20.00%) high severe
html/minify/document_fragment/github                                                                           
                        time:   [19.069 ms 19.221 ms 19.418 ms]
                        change: [-1.1241% -0.1140% +0.8650%] (p = 0.84 > 0.05)
                        No change in performance detected.
Found 2 outliers among 10 measurements (20.00%)
  2 (20.00%) high severe
html/minify/document_fragment/stackoverflow                                                                           
                        time:   [16.899 ms 16.921 ms 16.946 ms]
                        change: [-0.7586% -0.0360% +0.8779%] (p = 0.94 > 0.05)
                        No change in performance detected.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high severe
```
  
After using `swc_html_minify` to bench:
```
html/minify/document/css_spec                                                                           
                        time:   [41.873 ms 41.977 ms 42.118 ms]
                        change: [-11.739% -7.5798% -3.4009%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high severe
html/minify/document/github                                                                           
                        time:   [18.446 ms 18.492 ms 18.531 ms]
                        change: [-5.4652% -4.4001% -3.1595%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high severe
html/minify/document/stackoverflow                                                                           
                        time:   [14.923 ms 14.947 ms 14.971 ms]
                        change: [-8.9518% -7.9947% -6.9481%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high severe

html/minify/document_fragment/css_spec                                                                           
                        time:   [40.270 ms 40.437 ms 40.610 ms]
                        change: [-10.855% -9.3940% -8.0995%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
html/minify/document_fragment/github                                                                           
                        time:   [17.962 ms 17.991 ms 18.038 ms]
                        change: [-5.2539% -4.4223% -3.7342%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
html/minify/document_fragment/stackoverflow                                                                           
                        time:   [14.734 ms 14.760 ms 14.796 ms]
                        change: [-9.1301% -8.4154% -7.6186%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high severe
```

**BREAKING CHANGE:**

Yes

**Related issue (if exists):**

No